[refactor(react)] 重構navbar
